Move notifySubscriber back to EventService
