+++ events/quarkus/src/main/java/org/projectnessie/events/quarkus/delivery/RetriableEventDelivery.java
@@ -17,7 +17,7 @@ package org.projectnessie.events.quarkus.delivery;
 
 import java.time.Duration;
 import org.projectnessie.events.api.Event;
-import org.projectnessie.events.service.util.SubscriberUtil;
+import org.projectnessie.events.service.EventService;
 import org.projectnessie.events.spi.EventSubscriber;
 
 /**
@@ -57,7 +57,7 @@ public abstract class RetriableEventDelivery implements EventDelivery {
   /**
    * Tries to deliver the event to the subscriber.
    *
-   * @implSpec Implementers should call {@link SubscriberUtil#notifySubscriber(EventSubscriber,
+   * @implSpec Implementers should call {@link EventService#notifySubscriber(EventSubscriber,
    *     Event)}.
    * @param currentAttempt The current delivery attempt, starting at 1.
    */
